requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
httpx
google-cloud-storage
